move everything into a correct_lc function
